feat(datepicker): reuse logic with time picker
