+++ packages/react-magma-dom/src/components/DatePicker/DateField/StyledDateFieldInput.ts
@@ -1,5 +1,4 @@
 import styled from '@emotion/styled';
-import { transparentize } from 'polished';
 
 import { ThemeInterface } from '../../../theme/magma';
 import { inputWrapperStyles } from '../../InputBase';
@@ -16,32 +15,6 @@ export const InputsContainer = styled.div<{
   width: 100%;
   font-family: ${props => props.theme.bodyFont};
   position: relative;
-`;
-
-const getDividerColor = (
-  isInverse: boolean,
-  isFocused: boolean,
-  theme: ThemeInterface
-): string => {
-  if (isInverse) {
-    return isFocused
-      ? theme.colors.neutral100
-      : transparentize(0.3, theme.colors.neutral100);
-  }
-
-  return isFocused ? theme.colors.neutral700 : theme.colors.neutral500;
-};
-
-export const Divider = styled.span<{
-  theme: ThemeInterface;
-  isInverse?: boolean;
-  isFocused?: boolean;
-}>`
-  display: inline-block;
-  position: relative;
-  top: ${props => `-${props.theme.spaceScale.spacing01}`};
-  color: ${props =>
-    getDividerColor(props.isInverse, props.isFocused, props.theme)};
 `;
 
 export const IconWrapper = styled.div<{ theme: ThemeInterface }>`
