+++ packages/react-magma-dom/src/utils/index.ts
@@ -356,10 +356,6 @@ export const mergeRefs = <T>(...refs: Array<React.Ref<T> | undefined>) => {
   };
 };
 
-export function isNotEmpty(string: string | null | undefined): boolean {
-  return string !== null && string !== undefined && string.trim() !== '';
-}
-
 export const handleNumericBeforeInput = (
   e: React.FormEvent<HTMLInputElement>
 ) => {
